+++ examples/advanced/fedot_based_solutions/graph_model_optimization.py
@@ -17,8 +17,7 @@ from fedot.core.optimisers.graph import OptGraph, OptNode
 from fedot.core.optimisers.optimizer import GraphGenerationParams
 from fedot.core.pipelines.convert import graph_structure_as_nx_graph
 from fedot.core.utils import fedot_project_root
-from fedot.core.validation.objective import Objective
-from fedot.core.validation.objective_eval import ObjectiveEvaluate
+from fedot.core.optimisers.objective import Objective, ObjectiveEvaluate
 
 random.seed(1)
 np.random.seed(1)
