+++ fedot/core/optimisers/objective/__init__.py
@@ -1,0 +1,4 @@
+from .objective import Objective, ObjectiveFunction
+from .objective_eval import ObjectiveEvaluate
+from .data_objective_eval import DataObjectiveEvaluate, DataSource
+from .data_objective_builder import DataObjectiveBuilder
