+++ fedot/core/optimisers/objective/data_objective_eval.py
@@ -1,64 +1,19 @@
-from abc import abstractmethod, ABC
 from datetime import timedelta
-from typing import List, Iterable, Tuple, Callable, Optional, Sequence, TypeVar, Generic
+from typing import Optional, Callable, Iterable, Tuple
 
 import numpy as np
 
 from fedot.core.composer.cache import OperationsCache
-from fedot.core.dag.graph import Graph
 from fedot.core.data.data import InputData
 from fedot.core.log import Log, default_log
 from fedot.core.operations.model import Model
-from fedot.core.optimisers.fitness import Fitness
 from fedot.core.pipelines.pipeline import Pipeline
-from fedot.core.validation.objective import Objective, to_fitness
+from fedot.core.optimisers.fitness import Fitness
+from .objective import Objective, to_fitness
+from .objective_eval import ObjectiveEvaluate
 
 
 DataSource = Callable[[], Iterable[Tuple[InputData, InputData]]]
-
-
-G = TypeVar('G', bound=Graph, covariant=True)
-
-
-class ObjectiveEvaluate(ABC, Generic[G]):
-    """Defines how Objective must be evaluated on Graphs.
-
-     Responsibilities:
-     - Graph-specific evaluation policy: typically, Graphs require some kind of evaluation
-     before Objective could be estimated on them. E.g. Machine-learning pipelines must be
-     fit on train data before they could be evaluated on the test data.
-     - Objective-specific estimation: typically objectives require additional parameters
-     besides Graphs for estimation, e.g. test data for estimation of prediction quality.
-     - Optionally, compute additional statistics for Graphs (intermediate metrics).
-
-     Default implementation is just a closure that calls :param objective: with
-      redirected keyword arguments :param objective_kwargs:
-    """
-
-    def __init__(self, objective: Objective, **objective_kwargs):
-        self._objective = objective
-        self._objective_kwargs = objective_kwargs
-
-    @property
-    def objective(self) -> Objective:
-        """Returns underlying objective."""
-        return self._objective
-
-    def __call__(self, graph: G) -> Fitness:
-        """Provides functional interface for ObjectiveEvaluate."""
-        return self.evaluate(graph)
-
-    def evaluate(self, graph: G) -> Fitness:
-        """Evaluate graph and compute its fitness."""
-        return self._objective(graph, **self._objective_kwargs)
-
-    def evaluate_intermediate_metrics(self, graph: G):
-        """Compute intermediate metrics for each graph node and store it there."""
-        pass
-
-    def cleanup(self, graph: G):
-        """Clean resources after graph evaluation, if necessary."""
-        pass
 
 
 class DataObjectiveEvaluate(ObjectiveEvaluate[Pipeline]):
