+++ fedot/core/optimisers/objective/objective.py
